Transformed height and width attributes to double #42

+++ src/main/java/com/klai/stl/domain/Photo.java
@@ -44,11 +44,11 @@ public class Photo implements Serializable {
 
     @NotNull
     @Column(name = "height", nullable = false)
-    private Integer height;
+    private Double height;
 
     @NotNull
     @Column(name = "width", nullable = false)
-    private Integer width;
+    private Double width;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "orientation")
@@ -149,29 +149,29 @@ public class Photo implements Serializable {
         this.order = order;
     }
 
-    public Integer getHeight() {
+    public Double getHeight() {
         return this.height;
     }
 
-    public Photo height(Integer height) {
+    public Photo height(Double height) {
         this.height = height;
         return this;
     }
 
-    public void setHeight(Integer height) {
+    public void setHeight(Double height) {
         this.height = height;
     }
 
-    public Integer getWidth() {
+    public Double getWidth() {
         return this.width;
     }
 
-    public Photo width(Integer width) {
+    public Photo width(Double width) {
         this.width = width;
         return this;
     }
 
-    public void setWidth(Integer width) {
+    public void setWidth(Double width) {
         this.width = width;
     }
